feat(security): 添加关于页面拦截器并优化登录逻辑
- 新增 AboutInterceptor 拦截器，用于保护关于页面
-优化用户登录逻辑，增加登录失败次数限制
- 移除不必要的拦截器类
- 更新静态页面中的导航链接

+++ src/main/java/com/jiang/mall/domain/entity/Order.java
@@ -58,12 +58,12 @@ public class Order implements Serializable {
     /**
      * 订单状态
      */
-    private Integer status;
+    private byte status;
 
     /**
      * 支付方式
      */
-    private Integer paymentMethod;
+    private byte paymentMethod;
 
     /**
      * 创建时间，自动填充
